% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{cmapply}
\alias{cmapply}
\title{mapply after concatenating input data}
\usage{
cmapply(fn, ...)
}
\arguments{
\item{fn}{function}

\item{...}{all remaining mapply parameters}
}
\description{
mapply after concatenating input data
}
