% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{coltypes}
\alias{coltypes}
\title{Returns the column types for a data.table}
\usage{
coltypes(data)
}
\arguments{
\item{data}{- a data.table}
}
\value{
column types data.table with columns:
    variable  - name of the variable
    type      - type of the variable
    is_factor - TRUE/FALSE if the variable column is a factor
}
\description{
Returns the column types for a data.table
}
\examples{
    coltypes(data.table(iris))
    coltypes(data.table(iris))[type=="double"]

}
