% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{melt_all}
\alias{melt_all}
\title{Same as melt but melts all columns in the data.table}
\usage{
melt_all(data, variable.factor = F, ...)
}
\arguments{
\item{data}{- data.table}

\item{variable.factor}{(default=FALSE) - same as in data.table but with new default}

\item{...}{- all other params of the}
}
\description{
Same as melt but melts all columns in the data.table
}
