% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{silent.as.numeric}
\alias{silent.as.numeric}
\title{as.numeric but with suppressed warnings function}
\usage{
silent.as.numeric(x)
}
\arguments{
\item{value}{input value}
}
\value{
value cast to numeric type
}
\description{
as.numeric but with suppressed warnings function
}
