% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{vector.to.data.table}
\alias{vector.to.data.table}
\title{Converts a vector to data.table}
\usage{
vector.to.data.table(vec)
}
\arguments{
\item{vec}{vector to be converted (can have names).}
}
\value{
data.table with 1 row and cols named V1,...Vn for unnamed vectors or the names of the vector entries as column names
}
\description{
If the vector has names, such names are passed
to the output data.table's column names.
}
